Combine root/store & root/root_store
